+++ src/requisition-view/full-supply.routes.js
@@ -34,16 +34,17 @@
             resolve: {
                 lineItems: function(paginationService, requisition, $stateParams, $filter, requisitionValidator) {
 					return paginationService.registerList(requisitionValidator.isLineItemValid, $stateParams, function() {
-                        var filterObject = requisition.template.hideSkippedLineItems() ? {
-                            skipped: "!true",
-                            $program: {
-                                fullSupply: true
-                            }
-                        } : {
-                            $program: {
-                                fullSupply: true
-                            }
-                        };
+                        var filterObject = requisition.template.hideSkippedLineItems() ?
+                                            {
+                                                skipped: "!true",
+                                                $program: {
+                                                    fullSupply: true
+                                                }
+                                            } : {
+                                                $program: {
+                                                    fullSupply: true
+                                                }
+                                            };
                         var fullSupplyLineItems = $filter('filter')(requisition.requisitionLineItems, filterObject);
 
                         return $filter('orderBy')(fullSupplyLineItems, [