+++ CHANGELOG.md
@@ -10,6 +10,10 @@ New functionality added in a backwards-compatible manner:
 * [OLMIS-4748](https://openlmis.atlassian.net/browse/OLMIS-4748): Added disabling Total Losses and Adjustments modal for Stock Based Requisition.
 * [OLMIS-4747](https://openlmis.atlassian.net/browse/OLMIS-4747): Added Total Received Quantity and Total Consumed Quantity to the stock based columns.
 * [OLMIS-4760](https://openlmis.atlassian.net/browse/OLMIS-4760): Added Total Stockout Days to the stock based columns.
+* [OLMIS-4684](https://openlmis.atlassian.net/browse/OLMIS-4684): Hide skipped full supply products from create requisition when template is configured to hide.
+* [OLMIS-4685](https://openlmis.atlassian.net/browse/OLMIS-4685): Hide skipped full supply products from approve requisition when template is configured to hide.
+* [OLMIS-4686](https://openlmis.atlassian.net/browse/OLMIS-4686): Added an add product button and modal window to un-skip products
+* [OLMIS-4686](https://openlmis.atlassian.net/browse/OLMIS-4687): Allow users to search from skipped line items. 
 
 Bug fixes:
 * [OLMIS-4509](https://openlmis.atlassian.net/browse/OLMIS-4509): Fixed redirect to 404 when clicking proceed for requisition while offline